Create a select input that scales to very large choice sets

A drop-in replacement for shiny::selectInput() built on shinyWidgets::virtualSelectInput(). The underlying virtual-select library renders only the visible slice of the choice list, so a column with tens of thousands of unique values stays responsive and readable instead of producing an unusable wall of options.

Usage

viz_select_input(
  inputId,
  label,
  choices,
  selected = NULL,
  multiple = FALSE,
  search = NULL,
  width = "100%",
  ...
)

Arguments

inputId

The input slot that will be used to access the value.

label

Display label for the control, or NULL for no label.

choices

A vector or named list of values to select from, in the same form accepted by shiny::selectInput().

selected

The initially selected value(s). Defaults to the first choice for single selects and nothing for multiple selects.

multiple

Logical, whether multiple values can be selected.

search

Logical, whether to show a search box. Defaults to NULL, which enables the search box once there are more than 10 choices.

width

The width of the input, e.g. "100%" or "400px".

...

Further arguments passed to shinyWidgets::virtualSelectInput(), including any raw virtual-select property such as optionsCount or zIndex.

Value

A shiny.tag object to be included in a UI definition.

Details

Defaults are chosen so this behaves like shiny::selectInput(): when selected is not supplied and multiple = FALSE, the first choice is selected. An empty-string choice (the "no selection" convention used throughout this package) is relabelled "(none)" so it is visible in the dropdown, while its value remains "".

See also

Author

Jared Andrews

Examples

library(VizModules)
viz_select_input("gene", "Gene", choices = c("", paste0("GENE", 1:1000)))
#> <div class="form-group shiny-input-container" style="width:100%;">
#>   <label class="control-label" id="gene-label" for="gene">Gene</label>
#>   <div id="gene" class="virtual-select" style="width:100%;max-width:none;display:block;" data-update="change">
#>     <script type="application/json" data-for="gene">{"stateInput":false,"options":{"type":["transpose"],"choices":{"label":["(none)","GENE1","GENE2","GENE3","GENE4","GENE5","GENE6","GENE7","GENE8","GENE9","GENE10","GENE11","GENE12","GENE13","GENE14","GENE15","GENE16","GENE17","GENE18","GENE19","GENE20","GENE21","GENE22","GENE23","GENE24","GENE25","GENE26","GENE27","GENE28","GENE29","GENE30","GENE31","GENE32","GENE33","GENE34","GENE35","GENE36","GENE37","GENE38","GENE39","GENE40","GENE41","GENE42","GENE43","GENE44","GENE45","GENE46","GENE47","GENE48","GENE49","GENE50","GENE51","GENE52","GENE53","GENE54","GENE55","GENE56","GENE57","GENE58","GENE59","GENE60","GENE61","GENE62","GENE63","GENE64","GENE65","GENE66","GENE67","GENE68","GENE69","GENE70","GENE71","GENE72","GENE73","GENE74","GENE75","GENE76","GENE77","GENE78","GENE79","GENE80","GENE81","GENE82","GENE83","GENE84","GENE85","GENE86","GENE87","GENE88","GENE89","GENE90","GENE91","GENE92","GENE93","GENE94","GENE95","GENE96","GENE97","GENE98","GENE99","GENE100","GENE101","GENE102","GENE103","GENE104","GENE105","GENE106","GENE107","GENE108","GENE109","GENE110","GENE111","GENE112","GENE113","GENE114","GENE115","GENE116","GENE117","GENE118","GENE119","GENE120","GENE121","GENE122","GENE123","GENE124","GENE125","GENE126","GENE127","GENE128","GENE129","GENE130","GENE131","GENE132","GENE133","GENE134","GENE135","GENE136","GENE137","GENE138","GENE139","GENE140","GENE141","GENE142","GENE143","GENE144","GENE145","GENE146","GENE147","GENE148","GENE149","GENE150","GENE151","GENE152","GENE153","GENE154","GENE155","GENE156","GENE157","GENE158","GENE159","GENE160","GENE161","GENE162","GENE163","GENE164","GENE165","GENE166","GENE167","GENE168","GENE169","GENE170","GENE171","GENE172","GENE173","GENE174","GENE175","GENE176","GENE177","GENE178","GENE179","GENE180","GENE181","GENE182","GENE183","GENE184","GENE185","GENE186","GENE187","GENE188","GENE189","GENE190","GENE191","GENE192","GENE193","GENE194","GENE195","GENE196","GENE197","GENE198","GENE199","GENE200","GENE201","GENE202","GENE203","GENE204","GENE205","GENE206","GENE207","GENE208","GENE209","GENE210","GENE211","GENE212","GENE213","GENE214","GENE215","GENE216","GENE217","GENE218","GENE219","GENE220","GENE221","GENE222","GENE223","GENE224","GENE225","GENE226","GENE227","GENE228","GENE229","GENE230","GENE231","GENE232","GENE233","GENE234","GENE235","GENE236","GENE237","GENE238","GENE239","GENE240","GENE241","GENE242","GENE243","GENE244","GENE245","GENE246","GENE247","GENE248","GENE249","GENE250","GENE251","GENE252","GENE253","GENE254","GENE255","GENE256","GENE257","GENE258","GENE259","GENE260","GENE261","GENE262","GENE263","GENE264","GENE265","GENE266","GENE267","GENE268","GENE269","GENE270","GENE271","GENE272","GENE273","GENE274","GENE275","GENE276","GENE277","GENE278","GENE279","GENE280","GENE281","GENE282","GENE283","GENE284","GENE285","GENE286","GENE287","GENE288","GENE289","GENE290","GENE291","GENE292","GENE293","GENE294","GENE295","GENE296","GENE297","GENE298","GENE299","GENE300","GENE301","GENE302","GENE303","GENE304","GENE305","GENE306","GENE307","GENE308","GENE309","GENE310","GENE311","GENE312","GENE313","GENE314","GENE315","GENE316","GENE317","GENE318","GENE319","GENE320","GENE321","GENE322","GENE323","GENE324","GENE325","GENE326","GENE327","GENE328","GENE329","GENE330","GENE331","GENE332","GENE333","GENE334","GENE335","GENE336","GENE337","GENE338","GENE339","GENE340","GENE341","GENE342","GENE343","GENE344","GENE345","GENE346","GENE347","GENE348","GENE349","GENE350","GENE351","GENE352","GENE353","GENE354","GENE355","GENE356","GENE357","GENE358","GENE359","GENE360","GENE361","GENE362","GENE363","GENE364","GENE365","GENE366","GENE367","GENE368","GENE369","GENE370","GENE371","GENE372","GENE373","GENE374","GENE375","GENE376","GENE377","GENE378","GENE379","GENE380","GENE381","GENE382","GENE383","GENE384","GENE385","GENE386","GENE387","GENE388","GENE389","GENE390","GENE391","GENE392","GENE393","GENE394","GENE395","GENE396","GENE397","GENE398","GENE399","GENE400","GENE401","GENE402","GENE403","GENE404","GENE405","GENE406","GENE407","GENE408","GENE409","GENE410","GENE411","GENE412","GENE413","GENE414","GENE415","GENE416","GENE417","GENE418","GENE419","GENE420","GENE421","GENE422","GENE423","GENE424","GENE425","GENE426","GENE427","GENE428","GENE429","GENE430","GENE431","GENE432","GENE433","GENE434","GENE435","GENE436","GENE437","GENE438","GENE439","GENE440","GENE441","GENE442","GENE443","GENE444","GENE445","GENE446","GENE447","GENE448","GENE449","GENE450","GENE451","GENE452","GENE453","GENE454","GENE455","GENE456","GENE457","GENE458","GENE459","GENE460","GENE461","GENE462","GENE463","GENE464","GENE465","GENE466","GENE467","GENE468","GENE469","GENE470","GENE471","GENE472","GENE473","GENE474","GENE475","GENE476","GENE477","GENE478","GENE479","GENE480","GENE481","GENE482","GENE483","GENE484","GENE485","GENE486","GENE487","GENE488","GENE489","GENE490","GENE491","GENE492","GENE493","GENE494","GENE495","GENE496","GENE497","GENE498","GENE499","GENE500","GENE501","GENE502","GENE503","GENE504","GENE505","GENE506","GENE507","GENE508","GENE509","GENE510","GENE511","GENE512","GENE513","GENE514","GENE515","GENE516","GENE517","GENE518","GENE519","GENE520","GENE521","GENE522","GENE523","GENE524","GENE525","GENE526","GENE527","GENE528","GENE529","GENE530","GENE531","GENE532","GENE533","GENE534","GENE535","GENE536","GENE537","GENE538","GENE539","GENE540","GENE541","GENE542","GENE543","GENE544","GENE545","GENE546","GENE547","GENE548","GENE549","GENE550","GENE551","GENE552","GENE553","GENE554","GENE555","GENE556","GENE557","GENE558","GENE559","GENE560","GENE561","GENE562","GENE563","GENE564","GENE565","GENE566","GENE567","GENE568","GENE569","GENE570","GENE571","GENE572","GENE573","GENE574","GENE575","GENE576","GENE577","GENE578","GENE579","GENE580","GENE581","GENE582","GENE583","GENE584","GENE585","GENE586","GENE587","GENE588","GENE589","GENE590","GENE591","GENE592","GENE593","GENE594","GENE595","GENE596","GENE597","GENE598","GENE599","GENE600","GENE601","GENE602","GENE603","GENE604","GENE605","GENE606","GENE607","GENE608","GENE609","GENE610","GENE611","GENE612","GENE613","GENE614","GENE615","GENE616","GENE617","GENE618","GENE619","GENE620","GENE621","GENE622","GENE623","GENE624","GENE625","GENE626","GENE627","GENE628","GENE629","GENE630","GENE631","GENE632","GENE633","GENE634","GENE635","GENE636","GENE637","GENE638","GENE639","GENE640","GENE641","GENE642","GENE643","GENE644","GENE645","GENE646","GENE647","GENE648","GENE649","GENE650","GENE651","GENE652","GENE653","GENE654","GENE655","GENE656","GENE657","GENE658","GENE659","GENE660","GENE661","GENE662","GENE663","GENE664","GENE665","GENE666","GENE667","GENE668","GENE669","GENE670","GENE671","GENE672","GENE673","GENE674","GENE675","GENE676","GENE677","GENE678","GENE679","GENE680","GENE681","GENE682","GENE683","GENE684","GENE685","GENE686","GENE687","GENE688","GENE689","GENE690","GENE691","GENE692","GENE693","GENE694","GENE695","GENE696","GENE697","GENE698","GENE699","GENE700","GENE701","GENE702","GENE703","GENE704","GENE705","GENE706","GENE707","GENE708","GENE709","GENE710","GENE711","GENE712","GENE713","GENE714","GENE715","GENE716","GENE717","GENE718","GENE719","GENE720","GENE721","GENE722","GENE723","GENE724","GENE725","GENE726","GENE727","GENE728","GENE729","GENE730","GENE731","GENE732","GENE733","GENE734","GENE735","GENE736","GENE737","GENE738","GENE739","GENE740","GENE741","GENE742","GENE743","GENE744","GENE745","GENE746","GENE747","GENE748","GENE749","GENE750","GENE751","GENE752","GENE753","GENE754","GENE755","GENE756","GENE757","GENE758","GENE759","GENE760","GENE761","GENE762","GENE763","GENE764","GENE765","GENE766","GENE767","GENE768","GENE769","GENE770","GENE771","GENE772","GENE773","GENE774","GENE775","GENE776","GENE777","GENE778","GENE779","GENE780","GENE781","GENE782","GENE783","GENE784","GENE785","GENE786","GENE787","GENE788","GENE789","GENE790","GENE791","GENE792","GENE793","GENE794","GENE795","GENE796","GENE797","GENE798","GENE799","GENE800","GENE801","GENE802","GENE803","GENE804","GENE805","GENE806","GENE807","GENE808","GENE809","GENE810","GENE811","GENE812","GENE813","GENE814","GENE815","GENE816","GENE817","GENE818","GENE819","GENE820","GENE821","GENE822","GENE823","GENE824","GENE825","GENE826","GENE827","GENE828","GENE829","GENE830","GENE831","GENE832","GENE833","GENE834","GENE835","GENE836","GENE837","GENE838","GENE839","GENE840","GENE841","GENE842","GENE843","GENE844","GENE845","GENE846","GENE847","GENE848","GENE849","GENE850","GENE851","GENE852","GENE853","GENE854","GENE855","GENE856","GENE857","GENE858","GENE859","GENE860","GENE861","GENE862","GENE863","GENE864","GENE865","GENE866","GENE867","GENE868","GENE869","GENE870","GENE871","GENE872","GENE873","GENE874","GENE875","GENE876","GENE877","GENE878","GENE879","GENE880","GENE881","GENE882","GENE883","GENE884","GENE885","GENE886","GENE887","GENE888","GENE889","GENE890","GENE891","GENE892","GENE893","GENE894","GENE895","GENE896","GENE897","GENE898","GENE899","GENE900","GENE901","GENE902","GENE903","GENE904","GENE905","GENE906","GENE907","GENE908","GENE909","GENE910","GENE911","GENE912","GENE913","GENE914","GENE915","GENE916","GENE917","GENE918","GENE919","GENE920","GENE921","GENE922","GENE923","GENE924","GENE925","GENE926","GENE927","GENE928","GENE929","GENE930","GENE931","GENE932","GENE933","GENE934","GENE935","GENE936","GENE937","GENE938","GENE939","GENE940","GENE941","GENE942","GENE943","GENE944","GENE945","GENE946","GENE947","GENE948","GENE949","GENE950","GENE951","GENE952","GENE953","GENE954","GENE955","GENE956","GENE957","GENE958","GENE959","GENE960","GENE961","GENE962","GENE963","GENE964","GENE965","GENE966","GENE967","GENE968","GENE969","GENE970","GENE971","GENE972","GENE973","GENE974","GENE975","GENE976","GENE977","GENE978","GENE979","GENE980","GENE981","GENE982","GENE983","GENE984","GENE985","GENE986","GENE987","GENE988","GENE989","GENE990","GENE991","GENE992","GENE993","GENE994","GENE995","GENE996","GENE997","GENE998","GENE999","GENE1000"],"value":["","GENE1","GENE2","GENE3","GENE4","GENE5","GENE6","GENE7","GENE8","GENE9","GENE10","GENE11","GENE12","GENE13","GENE14","GENE15","GENE16","GENE17","GENE18","GENE19","GENE20","GENE21","GENE22","GENE23","GENE24","GENE25","GENE26","GENE27","GENE28","GENE29","GENE30","GENE31","GENE32","GENE33","GENE34","GENE35","GENE36","GENE37","GENE38","GENE39","GENE40","GENE41","GENE42","GENE43","GENE44","GENE45","GENE46","GENE47","GENE48","GENE49","GENE50","GENE51","GENE52","GENE53","GENE54","GENE55","GENE56","GENE57","GENE58","GENE59","GENE60","GENE61","GENE62","GENE63","GENE64","GENE65","GENE66","GENE67","GENE68","GENE69","GENE70","GENE71","GENE72","GENE73","GENE74","GENE75","GENE76","GENE77","GENE78","GENE79","GENE80","GENE81","GENE82","GENE83","GENE84","GENE85","GENE86","GENE87","GENE88","GENE89","GENE90","GENE91","GENE92","GENE93","GENE94","GENE95","GENE96","GENE97","GENE98","GENE99","GENE100","GENE101","GENE102","GENE103","GENE104","GENE105","GENE106","GENE107","GENE108","GENE109","GENE110","GENE111","GENE112","GENE113","GENE114","GENE115","GENE116","GENE117","GENE118","GENE119","GENE120","GENE121","GENE122","GENE123","GENE124","GENE125","GENE126","GENE127","GENE128","GENE129","GENE130","GENE131","GENE132","GENE133","GENE134","GENE135","GENE136","GENE137","GENE138","GENE139","GENE140","GENE141","GENE142","GENE143","GENE144","GENE145","GENE146","GENE147","GENE148","GENE149","GENE150","GENE151","GENE152","GENE153","GENE154","GENE155","GENE156","GENE157","GENE158","GENE159","GENE160","GENE161","GENE162","GENE163","GENE164","GENE165","GENE166","GENE167","GENE168","GENE169","GENE170","GENE171","GENE172","GENE173","GENE174","GENE175","GENE176","GENE177","GENE178","GENE179","GENE180","GENE181","GENE182","GENE183","GENE184","GENE185","GENE186","GENE187","GENE188","GENE189","GENE190","GENE191","GENE192","GENE193","GENE194","GENE195","GENE196","GENE197","GENE198","GENE199","GENE200","GENE201","GENE202","GENE203","GENE204","GENE205","GENE206","GENE207","GENE208","GENE209","GENE210","GENE211","GENE212","GENE213","GENE214","GENE215","GENE216","GENE217","GENE218","GENE219","GENE220","GENE221","GENE222","GENE223","GENE224","GENE225","GENE226","GENE227","GENE228","GENE229","GENE230","GENE231","GENE232","GENE233","GENE234","GENE235","GENE236","GENE237","GENE238","GENE239","GENE240","GENE241","GENE242","GENE243","GENE244","GENE245","GENE246","GENE247","GENE248","GENE249","GENE250","GENE251","GENE252","GENE253","GENE254","GENE255","GENE256","GENE257","GENE258","GENE259","GENE260","GENE261","GENE262","GENE263","GENE264","GENE265","GENE266","GENE267","GENE268","GENE269","GENE270","GENE271","GENE272","GENE273","GENE274","GENE275","GENE276","GENE277","GENE278","GENE279","GENE280","GENE281","GENE282","GENE283","GENE284","GENE285","GENE286","GENE287","GENE288","GENE289","GENE290","GENE291","GENE292","GENE293","GENE294","GENE295","GENE296","GENE297","GENE298","GENE299","GENE300","GENE301","GENE302","GENE303","GENE304","GENE305","GENE306","GENE307","GENE308","GENE309","GENE310","GENE311","GENE312","GENE313","GENE314","GENE315","GENE316","GENE317","GENE318","GENE319","GENE320","GENE321","GENE322","GENE323","GENE324","GENE325","GENE326","GENE327","GENE328","GENE329","GENE330","GENE331","GENE332","GENE333","GENE334","GENE335","GENE336","GENE337","GENE338","GENE339","GENE340","GENE341","GENE342","GENE343","GENE344","GENE345","GENE346","GENE347","GENE348","GENE349","GENE350","GENE351","GENE352","GENE353","GENE354","GENE355","GENE356","GENE357","GENE358","GENE359","GENE360","GENE361","GENE362","GENE363","GENE364","GENE365","GENE366","GENE367","GENE368","GENE369","GENE370","GENE371","GENE372","GENE373","GENE374","GENE375","GENE376","GENE377","GENE378","GENE379","GENE380","GENE381","GENE382","GENE383","GENE384","GENE385","GENE386","GENE387","GENE388","GENE389","GENE390","GENE391","GENE392","GENE393","GENE394","GENE395","GENE396","GENE397","GENE398","GENE399","GENE400","GENE401","GENE402","GENE403","GENE404","GENE405","GENE406","GENE407","GENE408","GENE409","GENE410","GENE411","GENE412","GENE413","GENE414","GENE415","GENE416","GENE417","GENE418","GENE419","GENE420","GENE421","GENE422","GENE423","GENE424","GENE425","GENE426","GENE427","GENE428","GENE429","GENE430","GENE431","GENE432","GENE433","GENE434","GENE435","GENE436","GENE437","GENE438","GENE439","GENE440","GENE441","GENE442","GENE443","GENE444","GENE445","GENE446","GENE447","GENE448","GENE449","GENE450","GENE451","GENE452","GENE453","GENE454","GENE455","GENE456","GENE457","GENE458","GENE459","GENE460","GENE461","GENE462","GENE463","GENE464","GENE465","GENE466","GENE467","GENE468","GENE469","GENE470","GENE471","GENE472","GENE473","GENE474","GENE475","GENE476","GENE477","GENE478","GENE479","GENE480","GENE481","GENE482","GENE483","GENE484","GENE485","GENE486","GENE487","GENE488","GENE489","GENE490","GENE491","GENE492","GENE493","GENE494","GENE495","GENE496","GENE497","GENE498","GENE499","GENE500","GENE501","GENE502","GENE503","GENE504","GENE505","GENE506","GENE507","GENE508","GENE509","GENE510","GENE511","GENE512","GENE513","GENE514","GENE515","GENE516","GENE517","GENE518","GENE519","GENE520","GENE521","GENE522","GENE523","GENE524","GENE525","GENE526","GENE527","GENE528","GENE529","GENE530","GENE531","GENE532","GENE533","GENE534","GENE535","GENE536","GENE537","GENE538","GENE539","GENE540","GENE541","GENE542","GENE543","GENE544","GENE545","GENE546","GENE547","GENE548","GENE549","GENE550","GENE551","GENE552","GENE553","GENE554","GENE555","GENE556","GENE557","GENE558","GENE559","GENE560","GENE561","GENE562","GENE563","GENE564","GENE565","GENE566","GENE567","GENE568","GENE569","GENE570","GENE571","GENE572","GENE573","GENE574","GENE575","GENE576","GENE577","GENE578","GENE579","GENE580","GENE581","GENE582","GENE583","GENE584","GENE585","GENE586","GENE587","GENE588","GENE589","GENE590","GENE591","GENE592","GENE593","GENE594","GENE595","GENE596","GENE597","GENE598","GENE599","GENE600","GENE601","GENE602","GENE603","GENE604","GENE605","GENE606","GENE607","GENE608","GENE609","GENE610","GENE611","GENE612","GENE613","GENE614","GENE615","GENE616","GENE617","GENE618","GENE619","GENE620","GENE621","GENE622","GENE623","GENE624","GENE625","GENE626","GENE627","GENE628","GENE629","GENE630","GENE631","GENE632","GENE633","GENE634","GENE635","GENE636","GENE637","GENE638","GENE639","GENE640","GENE641","GENE642","GENE643","GENE644","GENE645","GENE646","GENE647","GENE648","GENE649","GENE650","GENE651","GENE652","GENE653","GENE654","GENE655","GENE656","GENE657","GENE658","GENE659","GENE660","GENE661","GENE662","GENE663","GENE664","GENE665","GENE666","GENE667","GENE668","GENE669","GENE670","GENE671","GENE672","GENE673","GENE674","GENE675","GENE676","GENE677","GENE678","GENE679","GENE680","GENE681","GENE682","GENE683","GENE684","GENE685","GENE686","GENE687","GENE688","GENE689","GENE690","GENE691","GENE692","GENE693","GENE694","GENE695","GENE696","GENE697","GENE698","GENE699","GENE700","GENE701","GENE702","GENE703","GENE704","GENE705","GENE706","GENE707","GENE708","GENE709","GENE710","GENE711","GENE712","GENE713","GENE714","GENE715","GENE716","GENE717","GENE718","GENE719","GENE720","GENE721","GENE722","GENE723","GENE724","GENE725","GENE726","GENE727","GENE728","GENE729","GENE730","GENE731","GENE732","GENE733","GENE734","GENE735","GENE736","GENE737","GENE738","GENE739","GENE740","GENE741","GENE742","GENE743","GENE744","GENE745","GENE746","GENE747","GENE748","GENE749","GENE750","GENE751","GENE752","GENE753","GENE754","GENE755","GENE756","GENE757","GENE758","GENE759","GENE760","GENE761","GENE762","GENE763","GENE764","GENE765","GENE766","GENE767","GENE768","GENE769","GENE770","GENE771","GENE772","GENE773","GENE774","GENE775","GENE776","GENE777","GENE778","GENE779","GENE780","GENE781","GENE782","GENE783","GENE784","GENE785","GENE786","GENE787","GENE788","GENE789","GENE790","GENE791","GENE792","GENE793","GENE794","GENE795","GENE796","GENE797","GENE798","GENE799","GENE800","GENE801","GENE802","GENE803","GENE804","GENE805","GENE806","GENE807","GENE808","GENE809","GENE810","GENE811","GENE812","GENE813","GENE814","GENE815","GENE816","GENE817","GENE818","GENE819","GENE820","GENE821","GENE822","GENE823","GENE824","GENE825","GENE826","GENE827","GENE828","GENE829","GENE830","GENE831","GENE832","GENE833","GENE834","GENE835","GENE836","GENE837","GENE838","GENE839","GENE840","GENE841","GENE842","GENE843","GENE844","GENE845","GENE846","GENE847","GENE848","GENE849","GENE850","GENE851","GENE852","GENE853","GENE854","GENE855","GENE856","GENE857","GENE858","GENE859","GENE860","GENE861","GENE862","GENE863","GENE864","GENE865","GENE866","GENE867","GENE868","GENE869","GENE870","GENE871","GENE872","GENE873","GENE874","GENE875","GENE876","GENE877","GENE878","GENE879","GENE880","GENE881","GENE882","GENE883","GENE884","GENE885","GENE886","GENE887","GENE888","GENE889","GENE890","GENE891","GENE892","GENE893","GENE894","GENE895","GENE896","GENE897","GENE898","GENE899","GENE900","GENE901","GENE902","GENE903","GENE904","GENE905","GENE906","GENE907","GENE908","GENE909","GENE910","GENE911","GENE912","GENE913","GENE914","GENE915","GENE916","GENE917","GENE918","GENE919","GENE920","GENE921","GENE922","GENE923","GENE924","GENE925","GENE926","GENE927","GENE928","GENE929","GENE930","GENE931","GENE932","GENE933","GENE934","GENE935","GENE936","GENE937","GENE938","GENE939","GENE940","GENE941","GENE942","GENE943","GENE944","GENE945","GENE946","GENE947","GENE948","GENE949","GENE950","GENE951","GENE952","GENE953","GENE954","GENE955","GENE956","GENE957","GENE958","GENE959","GENE960","GENE961","GENE962","GENE963","GENE964","GENE965","GENE966","GENE967","GENE968","GENE969","GENE970","GENE971","GENE972","GENE973","GENE974","GENE975","GENE976","GENE977","GENE978","GENE979","GENE980","GENE981","GENE982","GENE983","GENE984","GENE985","GENE986","GENE987","GENE988","GENE989","GENE990","GENE991","GENE992","GENE993","GENE994","GENE995","GENE996","GENE997","GENE998","GENE999","GENE1000"]}},"config":{"multiple":false,"search":true,"selectedValue":"","hideClearButton":true,"autoSelectFirstOption":false,"showSelectedOptionsFirst":false,"showValueAsTags":false,"optionsCount":10,"noOfDisplayValues":50,"allowNewOption":false,"disableSelectAll":true,"disableOptionGroupCheckbox":true,"disabled":false,"dropboxWrapper":"body","zIndex":1060}}</script>
#>   </div>
#> </div>